refactor(e2e): extract verify.go into verify/ subpackage with focused files
Split the monolithic verify.go (489 lines) into a dedicated verify/
subpackage with 10 focused files, each named to match the type it
contains:

- manifest.go: ManifestInfo + ParseManifestInfo
- verifier.go: ResourceVerifier interface, dispatch, kind maps
- kubectl.go: kubectl helper functions with retry/backoff
- namespace.go: NamespaceVerifier (Tier 1)
- workload.go: WorkloadVerifier (Tier 1)
- resource_existence.go: ResourceExistenceVerifier (Tier 1)
- operator.go: OperatorComponentVerifier (Tier 4)
- crd_workload.go: CRDWorkloadVerifier (Tier 3)
- helm.go: HelmComponentVerifier (Tier 2)
- generic.go: GenericVerifier (fallback)

harness.go updated to import the verify subpackage. README.md updated
to document the new directory layout. Pure structural refactor -- no
logic changes. Smoke-tested with 10 KubernetesNamespace scenarios on
kind (all GREEN).
