+++ apis/org/openmcf/provider/kubernetes/aa_e2e/verify/verifier.go
@@ -1,0 +1,135 @@
+package verify
+
+import (
+	"context"
+	"strings"
+)
+
+// ResourceVerifier knows how to verify a specific Kubernetes resource type.
+type ResourceVerifier interface {
+	VerifyExists(ctx context.Context, kubeconfig string) error
+	VerifyAbsent(ctx context.Context, kubeconfig string) error
+}
+
+// operatorKinds lists manifest kind values (lowercased) for operator/controller
+// components. Operators install CRD controllers that watch resources but typically
+// do not expose a Kubernetes Service. Verification checks namespace + running
+// pods only (no service requirement).
+var operatorKinds = map[string]bool{
+	"kubernetesperconamongooperator":    true,
+	"kubernetesperconamysqloperator":    true,
+	"kubernetesperconapostgresoperator": true,
+	"kubernetessolroperator":            true,
+	"kubernetesaltinityoperator":        true,
+	"kuberneteselasticoperator":         true,
+	"kubernetesstrimzikafkaoperator":    true,
+	"kuberneteszalandopostgresoperator": true,
+}
+
+// crdWorkloadKinds lists manifest kind values (lowercased) for Tier 3
+// operator-dependent components. These create Custom Resources (e.g.,
+// Zalando Postgresql, Strimzi Kafka, ECK Elasticsearch) that are
+// reconciled by their prerequisite operator into pods and services.
+// Verification checks namespace + running pods + at least one service.
+var crdWorkloadKinds = map[string]bool{
+	"kubernetespostgres":      true,
+	"kuberneteskafka":         true,
+	"kuberneteselasticsearch": true,
+	"kubernetesmongodb":       true,
+	"kubernetessolr":          true,
+	"kubernetesclickhouse":    true,
+}
+
+// helmTier2Kinds lists all manifest kind values (lowercased) for Helm-based
+// Kubernetes components (Tier 2) that deploy applications with Services.
+// These must match the CloudResourceKind enum names from cloud_resource_kind.proto
+// (case-insensitive via lowercasing).
+//
+// Historical hack manifests use inconsistent kind names (e.g., "RedisKubernetes"
+// instead of "KubernetesRedis"). E2E manifests must use the enum name. Both
+// conventions are included here so the verifier works with either.
+var helmTier2Kinds = map[string]bool{
+	"kubernetesredis":    true,
+	"kubernetesnats":     true,
+	"kubernetesgrafana":  true,
+	"kubernetesneo4j":    true,
+	"kubernetesopenbao":  true,
+	"kubernetesopenfga":  true,
+	"kubernetesjenkins":  true,
+	"kubernetestemporal": true,
+	"kubernetesargocd":   true,
+	"kubernetesharbor":   true,
+	"kubernetesgitlab":   true,
+	"kuberneteslocust":   true,
+	"kubernetessignoz":   true,
+	"kuberneteskeycloak": true,
+	// Legacy hack manifest kind names (lowercased)
+	"rediskubernetes":  true,
+	"harborkubernetes": true,
+	"locustkubernetes": true,
+	"signozkubernetes": true,
+}
+
+// GetVerifierFromManifest creates the appropriate verifier by parsing the manifest.
+func GetVerifierFromManifest(manifestPath string) (ResourceVerifier, error) {
+	info, err := ParseManifestInfo(manifestPath)
+	if err != nil {
+		return nil, err
+	}
+
+	component := strings.ToLower(info.Kind)
+
+	switch component {
+	case "kubernetesnamespace":
+		return &NamespaceVerifier{Name: info.Name}, nil
+
+	case "kubernetesdeployment":
+		return &WorkloadVerifier{
+			Namespace: info.Namespace,
+			Kind:      "deployment",
+			Name:      info.Name,
+		}, nil
+
+	case "kubernetesstatefulset":
+		return &WorkloadVerifier{
+			Namespace: info.Namespace,
+			Kind:      "statefulset",
+			Name:      info.Name,
+		}, nil
+
+	case "kubernetessecret":
+		return &ResourceExistenceVerifier{
+			Namespace: info.Namespace,
+			Kind:      "secret",
+			Name:      info.Name,
+		}, nil
+
+	case "kubernetesservice":
+		return &ResourceExistenceVerifier{
+			Namespace: info.Namespace,
+			Kind:      "service",
+			Name:      info.Name,
+		}, nil
+
+	default:
+		if operatorKinds[component] {
+			return &OperatorComponentVerifier{
+				Namespace:     info.Namespace,
+				ComponentName: info.Name,
+			}, nil
+		}
+		if crdWorkloadKinds[component] {
+			return &CRDWorkloadVerifier{
+				Namespace:     info.Namespace,
+				ComponentName: info.Name,
+			}, nil
+		}
+		if helmTier2Kinds[component] {
+			return &HelmComponentVerifier{
+				Namespace:     info.Namespace,
+				ComponentName: info.Name,
+			}, nil
+		}
+		return &GenericVerifier{Component: component}, nil
+	}
+}
